Refactor: Rename internal sort package to sorting
