+++ internal/sorting/sorting_test.go
@@ -1,4 +1,4 @@
-package sort
+package sorting
 
 import (
 	"fmt"
